Refactor Terms Settings Editor
