+++ public/app/plugins/datasource/elasticsearch/components/QueryEditor/BucketAggregationsEditor/SettingsEditor/index.tsx
@@ -1,14 +1,14 @@
-import { InlineField, Input, Select } from '@grafana/ui';
+import { InlineField, Input } from '@grafana/ui';
 import React, { ComponentProps } from 'react';
 import { useDispatch } from '../../../../hooks/useStatelessReducer';
 import { SettingsEditorContainer } from '../../SettingsEditorContainer';
 import { changeBucketAggregationSetting } from '../state/actions';
 import { BucketAggregation } from '../aggregations';
-import { bucketAggregationConfig, createOrderByOptionsFromMetrics, orderOptions, sizeOptions } from '../utils';
+import { bucketAggregationConfig } from '../utils';
 import { FiltersSettingsEditor } from './FiltersSettingsEditor';
 import { useDescription } from './useDescription';
-import { useQuery } from '../../ElasticsearchQueryContext';
 import { DateHistogramSettingsEditor } from './DateHistogramSettingsEditor';
+import { TermsSettingsEditor } from './TermsSettingsEditor';
 
 export const inlineFieldProps: Partial<ComponentProps<typeof InlineField>> = {
   labelWidth: 16,
@@ -21,59 +21,13 @@ interface Props {
 export const SettingsEditor = ({ bucketAgg }: Props) => {
   const dispatch = useDispatch();
 
-  const { metrics } = useQuery();
   const settingsDescription = useDescription(bucketAgg);
-  const orderBy = createOrderByOptionsFromMetrics(metrics);
 
   return (
     <SettingsEditorContainer label={settingsDescription}>
-      {bucketAgg.type === 'terms' && (
-        <>
-          <InlineField label="Order" {...inlineFieldProps}>
-            <Select
-              onChange={(e) => dispatch(changeBucketAggregationSetting(bucketAgg, 'order', e.value!))}
-              options={orderOptions}
-              value={bucketAgg.settings?.order || bucketAggregationConfig[bucketAgg.type].defaultSettings?.order}
-            />
-          </InlineField>
-
-          <InlineField label="Size" {...inlineFieldProps}>
-            <Select
-              onChange={(e) => dispatch(changeBucketAggregationSetting(bucketAgg, 'size', e.value!))}
-              options={sizeOptions}
-              value={bucketAgg.settings?.size || bucketAggregationConfig[bucketAgg.type].defaultSettings?.size}
-              allowCustomValue
-            />
-          </InlineField>
-
-          <InlineField label="Min Doc Count" {...inlineFieldProps}>
-            <Input
-              onBlur={(e) => dispatch(changeBucketAggregationSetting(bucketAgg, 'min_doc_count', e.target.value!))}
-              defaultValue={
-                bucketAgg.settings?.min_doc_count ||
-                bucketAggregationConfig[bucketAgg.type].defaultSettings?.min_doc_count
-              }
-            />
-          </InlineField>
-
-          <InlineField label="Order By" {...inlineFieldProps}>
-            <Select
-              onChange={(e) => dispatch(changeBucketAggregationSetting(bucketAgg, 'orderBy', e.value!))}
-              options={orderBy}
-              value={bucketAgg.settings?.orderBy || bucketAggregationConfig[bucketAgg.type].defaultSettings?.orderBy}
-            />
-          </InlineField>
-
-          <InlineField label="Missing" {...inlineFieldProps}>
-            <Input
-              onBlur={(e) => dispatch(changeBucketAggregationSetting(bucketAgg, 'missing', e.target.value!))}
-              defaultValue={
-                bucketAgg.settings?.missing || bucketAggregationConfig[bucketAgg.type].defaultSettings?.missing
-              }
-            />
-          </InlineField>
-        </>
-      )}
+      {bucketAgg.type === 'terms' && <TermsSettingsEditor bucketAgg={bucketAgg} />}
+      {bucketAgg.type === 'date_histogram' && <DateHistogramSettingsEditor bucketAgg={bucketAgg} />}
+      {bucketAgg.type === 'filters' && <FiltersSettingsEditor value={bucketAgg} />}
 
       {bucketAgg.type === 'geohash_grid' && (
         <InlineField label="Precision" {...inlineFieldProps}>
@@ -85,8 +39,6 @@ export const SettingsEditor = ({ bucketAgg }: Props) => {
           />
         </InlineField>
       )}
-
-      {bucketAgg.type === 'date_histogram' && <DateHistogramSettingsEditor bucketAgg={bucketAgg} />}
 
       {bucketAgg.type === 'histogram' && (
         <>
@@ -110,8 +62,6 @@ export const SettingsEditor = ({ bucketAgg }: Props) => {
           </InlineField>
         </>
       )}
-
-      {bucketAgg.type === 'filters' && <FiltersSettingsEditor value={bucketAgg} />}
     </SettingsEditorContainer>
   );
 };
